Refactor ep1
