+++ src/main/java/com/infoshareacademy/drinkers/ApplicationProperties.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 import java.io.IOException;
 import java.io.StringReader;
