+++ src/main/java/com/infoshareacademy/drinkers/ConsoleColors.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 public class ConsoleColors {
     // Reset
