+++ src/main/java/com/infoshareacademy/drinkers/DrinkIDComparator.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 import java.util.Comparator;
