+++ src/main/java/com/infoshareacademy/drinkers/Drinks.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 import java.util.List;
 import java.util.Objects;
