+++ src/main/java/com/infoshareacademy/drinkers/FilterList.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 import java.util.ArrayList;
 import java.util.List;
