+++ src/main/java/com/infoshareacademy/drinkers/GsonCreator.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
