+++ src/main/java/com/infoshareacademy/drinkers/InstructionsLanguage.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 public enum InstructionsLanguage {
     ES("ES"),
