+++ src/main/java/com/infoshareacademy/drinkers/SortDrinks.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 import org.apache.log4j.Logger;
