+++ src/main/java/com/infoshareacademy/drinkers/SortItems.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy;
+package com.infoshareacademy.drinkers;
 
 public enum SortItems {
     ID("ID"),
