+++ src/main/java/com/infoshareacademy/drinkers/service/DrinkBuilder.java
@@ -1,0 +1,4 @@
+package com.infoshareacademy.drinkers.service;
+
+public class DrinkBuilder {
+}
