--- src/main/java/service/DrinkBuilder.java
@@ -1,4 +1,0 @@
-package service;
-
-public class DrinkBuilder {
-}
